gpt_resume/requirements.txt
django
django-cors-headers
djangorestframework
openai
pymupdf
python-dotenv
